feat: negative number graph support

+++ upload.sh
@@ -2,13 +2,13 @@
 
 DEVICE=/dev/ttyUSB0
 
-if [ "$MAIN_ONLY" != "true" ]
+if [ "$MAIN_ONLY" != "true" ] && [ "$ONLY_MAIN" != "true" ]
 then
     echo Removing old code...
     rshell --port $DEVICE --quiet rm -r /pyboard/src
 fi
 echo Adding new code...
-if [ "$MAIN_ONLY" != "true" ]
+if [ "$MAIN_ONLY" != "true" ] && [ "$ONLY_MAIN" != "true" ]
 then
     ampy --port $DEVICE put src/
 else
